Change subprocess definition
